Task3: Add files with utility functions
