+++ task1/utility.h
@@ -1,0 +1,7 @@
+#pragma once
+
+// generates random number in range [low, high]
+int getRandomNumber(int low, int high);
+
+// get user number in range [low, high]
+int getUserNumber(int low, int high);
